move hidden form to components
